Django init and cleanup
I moved the mockup and the prototype files to the Old Assignment folder in an attempt to cleanup. I also created the Django project and just started to configure it a tad. There is a readme.txt file I created with instructions to starting up the project, for first time and every time after that.